requirements/base.txt
# Django
Django==3.1.4
djangorestframework==3.11
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0

# Database
mysqlclient==2.0.1

# Gunicorn
gunicorn==20.0.4

# Utilities
python-decouple==3.3
drf-yasg==1.20.0
django-simple-history==2.12.0
Pillow==8.1.0
django-storages==1.11.1
boto3==1.17.34
google-api-python-client==2.1.0
django-filter==2.4.0
sentry-sdk==1.0.0
six==1.15.0

# Static files
whitenoise==5.1.0
